Move config type creation from docker.utils functions to classes in docker.types
