+++ docker/models/networks.py
@@ -98,7 +98,7 @@ class NetworkCollection(Collection):
             driver (str): Name of the driver used to create the network
             options (dict): Driver options as a key-value dictionary
             ipam (dict): Optional custom IP scheme for the network.
-                Created with :py:meth:`~docker.utils.create_ipam_config`.
+                Created with :py:class:`~docker.types.IPAMConfig`.
             check_duplicate (bool): Request daemon to check for networks with
                 same name. Default: ``True``.
             internal (bool): Restrict external access to the network. Default
@@ -125,11 +125,11 @@ class NetworkCollection(Collection):
 
             .. code-block:: python
 
-                >>> ipam_pool = docker.utils.create_ipam_pool(
+                >>> ipam_pool = docker.types.IPAMPool(
                     subnet='192.168.52.0/24',
                     gateway='192.168.52.254'
                 )
-                >>> ipam_config = docker.utils.create_ipam_config(
+                >>> ipam_config = docker.types.IPAMConfig(
                     pool_configs=[ipam_pool]
                 )
                 >>> client.networks.create(
