+++ docker/types/__init__.py
@@ -1,8 +1,9 @@
 # flake8: noqa
-from .containers import LogConfig, Ulimit
+from .containers import ContainerConfig, HostConfig, LogConfig, Ulimit
+from .healthcheck import Healthcheck
+from .networks import EndpointConfig, IPAMConfig, IPAMPool, NetworkingConfig
 from .services import (
     ContainerSpec, DriverConfig, EndpointSpec, Mount, Resources, RestartPolicy,
     TaskTemplate, UpdateConfig
 )
-from .healthcheck import Healthcheck
 from .swarm import SwarmSpec, SwarmExternalCA
