+++ docker/utils/__init__.py
@@ -3,12 +3,9 @@ from .utils import (
     compare_version, convert_port_bindings, convert_volume_binds,
     mkbuildcontext, tar, exclude_paths, parse_repository_tag, parse_host,
     kwargs_from_env, convert_filters, datetime_to_timestamp,
-    create_host_config, create_container_config, parse_bytes, ping_registry,
-    parse_env_file, version_lt, version_gte, decode_json_header, split_command,
-    create_ipam_config, create_ipam_pool, parse_devices, normalize_links,
-    convert_service_networks,
+    create_host_config, parse_bytes, ping_registry, parse_env_file, version_lt,
+    version_gte, decode_json_header, split_command, create_ipam_config,
+    create_ipam_pool, parse_devices, normalize_links, convert_service_networks,
 )
 
-from ..types import LogConfig, Ulimit
-from ..types import SwarmExternalCA, SwarmSpec
 from .decorators import check_resource, minimum_version, update_headers
